pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "transgenderwachttijd"
version = "0.1.0"
description = "Website which displays the current waiting times for transgender health care in the Netherlands."
authors = ["Danielle Huisman <danielle@huisman.me>"]
maintainers = ["Danielle Huisman <danielle@huisman.me>"]
license = "MIT"
readme = "README.md"

homepage = "https://github.com/DanielleHuisman/transgenderwachttijd-backend"
repository = "https://github.com/DanielleHuisman/transgenderwachttijd-backend"

keywords = ["transgender", "waiting", "times"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "License :: OSI Approved :: MIT License",
    "Private :: Do Not Upload",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3"
]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.7"
django-cors-headers = "^4.3.0"
django-filter = "^23.3"
django-modeltranslation = "^0.18.11"
django-q = "^1.3.9"
feedparser = "^6.0.10"
gunicorn = "^21.2.0"
requests = "^2.31.0"
pdfreader = "^0.1.12"
mysqlclient = "^2.2.0"
python-dotenv = "^1.0.0"
pytz = "^2023.3.post1"
sentry-sdk = "^1.34.0"
strawberry-graphql = "^0.211.1"
strawberry-graphql-django = "^0.23.0"
beautifulsoup4 = "^4.12.2"
croniter = "^2.0.1"

[tool.poetry.extras]
django-q = ["sentry"]
